refactor(cloud)!: Use 'metro' as default vs. 'metros'
